+++ source/projects/py/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [0.3.x]
 
+- Added `api.Dictionary.get_atomarray` and `api.Dictionary.append_atomarray` to api
+
+- Added `api.Dictionary.get_object` and `api.Dictionary.append_object` to api
+
 - Added `DatabaseView` and `DatabaseResult` objects to `api` module
 
 - Changed `api.Linklist` to use `ptr` instead of `lst`, and other objects similarly.